+++ flask_fixture/__init__.py
@@ -1,3 +1,3 @@
-from flask_fixture.decorator_endpoint import endpoint
+from flask_fixture.decorators.decorator_endpoint import endpoint
 from flask_fixture.fixture import local_server_url
-from flask_fixture.decorator_config import config
+from flask_fixture.decorators.decorator_config import config
