--- flask_fixture/collection_of_routes.py
@@ -1,6 +1,0 @@
-from typing import List
-
-from flask_fixture.route_item import RouteItem
-
-
-routes: List[RouteItem] = []
