+++ flask_fixture/fixture.py
@@ -3,9 +3,9 @@ import multiprocessing
 import pytest
 
 from flask_fixture.runner import run_flask
-from flask_fixture.collection_of_routes import routes
-from flask_fixture.collection_of_importing_modules import modules
-from flask_fixture.collection_of_configs import configs
+from flask_fixture.state_storage.collection_of_routes import routes
+from flask_fixture.state_storage.collection_of_importing_modules import modules
+from flask_fixture.state_storage.collection_of_configs import configs
 from flask_fixture.errors import NotExpectedConfigFieldError, UnsuccessfulProcessStartupError
