+++ flask_fixture/state_storage/collection_of_configs.py
