+++ flask_fixture/state_storage/collection_of_routes.py
@@ -1,0 +1,6 @@
+from typing import List
+
+from flask_fixture.dataclasses.route_item import RouteItem
+
+
+routes: List[RouteItem] = []
